% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calculateIndex.R
\name{calculateIndex}
\alias{calculateIndex}
\title{Calculate a specified index}
\usage{
calculateIndex(normReflDF, indexDF, indexno)
}
\arguments{
\item{normReflDF}{A data frame of normal reflectances with three columns: location, wavelength, and normrefl (normal reflectance).}

\item{indexDF}{A data frame of indices that provides information on how to calculate each index.}

\item{indexno}{Index number}
}
\value{
A data frame of locations and calculated index values
}
\description{
Given a data frame of indices and a row number, calculate the specified index.
}

