% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calculateInterpNormRefl.R
\name{calculateInterpNormRefl}
\alias{calculateInterpNormRefl}
\title{Calculate interpolated normalized reflectance}
\usage{
calculateInterpNormRefl(eventdata, calreflavgs, interpolation, limitNR)
}
\arguments{
\item{eventdata}{A data frame of one observation made by a hyperspectrometer. This should have three columns: wavelength, irradiance, and radiance.}

\item{calreflavgs}{The reflectance averages from a calibration panel (usually averaged from multiple scans of a panel).}

\item{interpolation}{The interpolation type. Values include "linear", "natural" (spline), and "fmm" (cubic; the Forsythe, Malcolm and Moler method as described in the R help files for the spline() function. "...an exact cubic is fitted through the four points at each end of the data, and this is used to determine the end conditions")}

\item{limitNR}{The limit of the normalized reflectance that should be used. "nl" (no limit) = allow all values; "lim1" = if a value is greater than 1 or less than -1, change the value to 1 or -1, respectively; limNA" = if a value is greater than 1 or less than -1, change the value to NA.}
}
\value{
A data frame with four columns: wavelength, irradiance, radiance, and normrefl (normal reflectance).
}
\description{
Interpolate the radiance and irradiance data for a single event file. Then calculate normalized reflectance for the event file. Arguments include a single event dataset, the average panel reflectance, and the interpolation type (linear, spline, or cubic). Return a data frame with four columns: wavelength, irradiance, radiance, and normrefl (normal reflectance).
}

