\name{rHyperspec-package}
\alias{rHyperspec-package}
\alias{rHyperspec}
\docType{package}
\title{
\packageTitle{rHyperspec}
}
\description{
\packageDescription{rHyperspec}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rHyperspec}
\packageIndices{rHyperspec}
This is a description.
}
\author{
\packageAuthor{rHyperspec}

Maintainer: \packageMaintainer{rHyperspec}
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{

}
